LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/media-elements/networkState_during_progress-expected.txt

spec reference

  

FAIL audioElement.networkState should be NETWORK_LOADING during progress event assert_equals: expected 2 but got 1
FAIL videoElement.networkState should be NETWORK_LOADING during progress event assert_equals: expected 2 but got 1
